se cambia Long por Double en div

+++ src/main/java/co/com/sofka/pruebasMaven/calculator/BasicCalculator.java
@@ -26,7 +26,7 @@ public class BasicCalculator {
     }
 
     // Divide
-    public Long div(Long number1, Long number2) throws IllegalArgumentException {
+    public Double div(Long number1, Long number2) throws IllegalArgumentException {
         if(number2 == 0){
             //logger.error("cannot be divided by 0");
             throw new IllegalArgumentException("cannot be divided by 0");
